refactor(adapters): relocate report/html into adapters/driven/report (hexagonal P4 sweep)
Pure git mv of regis/report/html.py → regis/adapters/driven/report/html.py
(joining FileReportSink + the presentation renderer) + import-path updates in
the 6 importers; the vestigial regis/report/__init__.py is removed. html.py
imports nothing from regis, so the move is layering-trivial. No behaviour
change. Second subtree of the P4 physical sweep.
