--- regis/report/__init__.py
@@ -1,1 +1,0 @@
-"""Report renderers for regis."""
