+++ tests/report/test_html_verdict.py
@@ -1,4 +1,4 @@
-from regis.report.html import render_html_single
+from regis.adapters.driven.report.html import render_html_single
 
 
 def _report():
